feature(Map): added customView to marker.
